refactor Bond and Unbond
