+++ src/modals/JoinPool/Forms.tsx
@@ -8,7 +8,7 @@ import { IconProp } from '@fortawesome/fontawesome-svg-core';
 import { useModal } from 'contexts/Modal';
 import { useApi } from 'contexts/Api';
 import { useConnect } from 'contexts/Connect';
-import { BondFeedback } from 'library/Form/Bonding/BondFeedback';
+import { BondFeedback } from 'library/Form/Bond/BondFeedback';
 import { useSubmitExtrinsic } from 'library/Hooks/useSubmitExtrinsic';
 import { planckBnToUnit, unitToPlanckBn } from 'Utils';
 import { usePoolMembers } from 'contexts/Pools/PoolMembers';
