+++ src/modals/UpdateBond/Forms/BondSome.tsx
@@ -5,7 +5,7 @@ import { useState, useEffect } from 'react';
 import { useModal } from 'contexts/Modal';
 import { useApi } from 'contexts/Api';
 import { useConnect } from 'contexts/Connect';
-import { BondFeedback } from 'library/Form/Bonding/BondFeedback';
+import { BondFeedback } from 'library/Form/Bond/BondFeedback';
 import { useSubmitExtrinsic } from 'library/Hooks/useSubmitExtrinsic';
 import { planckBnToUnit } from 'Utils';
 import { EstimatedTxFee } from 'library/EstimatedTxFee';
