+++ src/pages/Nominate/Setup/Bond/index.tsx
@@ -4,7 +4,7 @@
 import { useState, useEffect } from 'react';
 import { useConnect } from 'contexts/Connect';
 import { useUi } from 'contexts/UI';
-import { BondFeedback } from 'library/Form/Bonding/BondFeedback';
+import { BondFeedback } from 'library/Form/Bond/BondFeedback';
 import { NominateStatusBar } from 'library/Form/NominateStatusBar';
 import { SetupStepProps } from 'library/SetupSteps/types';
 import { SetupType } from 'contexts/UI/types';
