+++ src/pages/Pools/Create/Bond/index.tsx
@@ -4,7 +4,7 @@
 import { useState, useEffect } from 'react';
 import { useConnect } from 'contexts/Connect';
 import { useUi } from 'contexts/UI';
-import { BondFeedback } from 'library/Form/Bonding/BondFeedback';
+import { BondFeedback } from 'library/Form/Bond/BondFeedback';
 import { CreatePoolStatusBar } from 'library/Form/CreatePoolStatusBar';
 import { SetupStepProps } from 'library/SetupSteps/types';
 import { SetupType } from 'contexts/UI/types';
